Add missing fixes
