+++ apiserver/apiserver.go
@@ -13,7 +13,7 @@ import (
 	swagger "github.com/swaggo/http-swagger"
 	"gocloud.dev/blob"
 	_ "gocloud.dev/blob/azureblob" // register azureblob driver
-	//_ "gocloud.dev/blob/fileblob"  // register fileblob driver
+	// _ "gocloud.dev/blob/fileblob"  // register fileblob driver
 	_ "gocloud.dev/blob/gcsblob" // register gcsblob driver
 	_ "gocloud.dev/blob/memblob" // register memblob driver
 	_ "gocloud.dev/blob/s3blob"  // register s3blob driver
